pyproject.toml
[project]
name = "quicktts"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "profanityfilter>=2.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.10",
    "yapper-tts>=0.7.0",
]
